Get the access token by creating routes
