+++ src/views/partials/header.hbs
@@ -1,5 +1,7 @@
 <header class="members-header container">
   <h1>{{title}}</h1>
   <p>{{description}}</p>
+  {{#if githubStatus}}
   <p>{{githubStatus}}</p>
+  {{/if}}
 </header>
